deployment/cloud_function/requirements.txt
google-cloud-aiplatform==1.38.0
google-cloud-bigquery==3.14.0
functions-framework==3.*
